requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
tqdm>=4.62.0
PyYAML>=6.0
matplotlib>=3.5.0
jupyter>=1.0.0
tensorboard>=2.9.0
scikit-learn>=1.0.0
Pillow>=8.3.0